LoginDialogButton Componentized
LoginDialogButton and HomePage are now separate components. At first I was going to separate the reducer and action files, but didn't want to bother complicating state management too much. Might be a change to make for the future.

While testing, I noticed some buggy behaviour. I am going to check if this was introduced, or if it was pre-existing.