Add Space for Listing Detailed View
This adds space on the right hand side so that we can display full listing content for when a listing is selected. Some styling changes have been made for the now-left-hand side of the screen so that it fits within its box. These changes aren't meant to be final. They were just so everything fit.