Added Search Tests, Started Advanced Search Test
Currently the two normal search tests are dependent on the content of the database. It should be somewhat easy for us to change this so that it reads the titles of two listings and then types in a keyword (and checks for it) based on those titles.

Advanced search has been started but is not complete.