Update creation and deletion test
This can be hooked in with other tests. I put creation/deletion into the same test because repeated login/logout takes a while (I'm fine if they are separated).

Email/pw should not be changed to the demo account. If that happens, it will delete the demo listings because it doesn't verify that it's deleting the same listing it created.